requirement.txt
aiogram==3.10.0
openai==1.3.7
python-dotenv==1.0.1
aiohttp==3.9.5


TELEGRAM_BOT_TOKEN =8274126714:AAHilTOIRtPlGSPQKKsF8eSRKIyrDJBxc_s


